added comments, renamed file
