+++ src/modules/core/components/UserPickerWithSearch/constants.ts
@@ -1,2 +1,2 @@
-export const DROPDOWN_HEIGHT = 180;
-export const DROPDOWN_ITEM_HEIGHT = 56;
+export const DROPDOWN_HEIGHT = 160;
+export const DROPDOWN_ITEM_HEIGHT = 32;
