+++ src/modules/dashboard/components/ExpenditurePage/Split/constants.ts
@@ -1,3 +1,6 @@
+// This is the initial recipient object used to create the initial value in the form.
+// It is also used as the initial value when you add a new recipient in the form.
+
 export const initalRecipient = {
   user: undefined,
   amount: { value: 0 },
